library: add basic IPC API
